+++ shiny/_versions.py
@@ -1,4 +1,4 @@
-shiny_html_deps = "1.13.0.9000"
+shiny_html_deps = "1.14.0.9000"
 bslib = "0.11.0.9000"
 htmltools = "0.5.9.9000"
 bootstrap = "5.3.8"